Refactor code for retaining formats in message-passing. (#2570)
